+++ BTrees/_base.py
@@ -942,6 +942,12 @@ class _Tree(_Base):
 
         removed_first_bucket, value = child._del(key)
 
+        # See comment in _set about small trees
+        if (len(data) == 1 and
+            child.__class__ is self._bucket_type and
+            child._p_oid is None):
+            self._p_changed = True
+
         # fix up the node key, but not for the 0'th one.
         if index > 0 and child.size and key == data[index].key:
             self._p_changed = True