+++ BTrees/tests/common.py
@@ -1146,6 +1146,37 @@ class BTreeTests(MappingBase):
         self.assertTrue(t._p_changed)
         self.assertEqual(t, t._p_jar.registered)
 
+    def testRemoveInSmallMapSetsChanged(self):
+        # A bug in the BTree Python implementation once caused
+        # deleting from a small btree to set _p_changed
+        t = self._makeOne()
+        # Note that for the property to actually hold, we have to fake a
+        # _p_jar and _p_oid
+        t._p_oid = b'\0\0\0\0\0'
+        class Jar(object):
+            def __init__(self):
+                self._cache = self
+                self.registered = None
+
+            def mru(self, arg):
+                pass
+            def readCurrent(self, arg):
+                pass
+            def register(self, arg):
+                self.registered = arg
+
+        t._p_jar = Jar()
+        t[0] = 1
+        # reset these, setting _firstbucket triggered a change
+        t._p_changed = False
+        t._p_jar.registered = None
+
+        # now remove the second value
+        del t[0]
+        self.assertTrue(t._p_changed)
+        self.assertEqual(t, t._p_jar.registered)
+
+
 class NormalSetTests(Base):
     # Test common to all set types
 
@@ -1353,6 +1384,37 @@ class NormalSetTests(Base):
             except StopIteration:
                 pass
             self.assertEqual(x, keys)
+
+    def testRemoveInSmallSetSetsChanged(self):
+        # A bug in the BTree TreeSet Python implementation once caused
+        # deleting an item in a small set to fail to set _p_changed
+        t = self._makeOne()
+        # Note that for the property to actually hold, we have to fake a
+        # _p_jar and _p_oid
+        t._p_oid = b'\0\0\0\0\0'
+        class Jar(object):
+            def __init__(self):
+                self._cache = self
+                self.registered = None
+
+            def mru(self, arg):
+                pass
+            def readCurrent(self, arg):
+                pass
+            def register(self, arg):
+                self.registered = arg
+
+        t._p_jar = Jar()
+        t.add(0)
+        # reset these, setting _firstbucket triggered a change
+        t._p_changed = False
+        t._p_jar.registered = None
+
+        # now remove the second value
+        t.remove(0)
+        self.assertTrue(t._p_changed)
+        self.assertEqual(t, t._p_jar.registered)
+
 
 class ExtendedSetTests(NormalSetTests):